refactor: move empty document check to helpers
